Добавил H2, swagger начал делать тесты

+++ src/main/java/com/alexpyslar03/productselectorbackend/service/ProductService.java
@@ -4,10 +4,9 @@ import com.alexpyslar03.productselectorbackend.dto.ProductDTO;
 import com.alexpyslar03.productselectorbackend.entity.Product;
 import com.alexpyslar03.productselectorbackend.repository.ProductRepository;
 import com.alexpyslar03.productselectorbackend.repository.RecipeRepository;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Set;
 
@@ -16,29 +15,43 @@ import java.util.Set;
 public class ProductService {
     private final ProductRepository productRepository;
     private final RecipeRepository recipeRepository;
+
     public Product create(ProductDTO dto) {
         return productRepository.save(Product.builder()
                 .name(dto.getName())
                 .image(dto.getImage())
-                .recipes(recipeRepository.findByIds(dto.getRecipeIds()))
+                .recipes(recipeRepository.findAllByIdIn(dto.getRecipeIds()))
                 .build());
     }
+
     public List<Product> readAll() {
         return productRepository.findAll();
     }
+
     public Product readById(Long id) {
-        return productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found - " + id));
+        return productRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Product not found - " + id));
     }
-    public List<Product> readByIds(List<Long> ids) {
-        return productRepository.findByIds(ids);
+
+    public Set<Product> readAllByIdIn(List<Long> ids) {
+        return productRepository.findAllByIdIn(ids);
     }
+
     public List<Product> readByRecipeId(Long id) {
         return productRepository.findByRecipesId(id);
     }
+
     public Product update(Product product) {
+        if (!productRepository.existsById(product.getId())) {
+            throw new RuntimeException("Product not found - " + product.getId());
+        }
         return productRepository.save(product);
     }
+
     public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new RuntimeException("Product not found - " + id);
+        }
         productRepository.deleteById(id);
     }
 }
